+++ src/parser/p_o_minus.rs
@@ -1,8 +1,8 @@
 use crate::parser::{Parselet, ParseError, parse_expression};
 use crate::lexer::Lexer;
 use std::rc::Rc;
-use crate::astplus::Expression;
-use crate::astplus::e_minus::{PrefixMinusExpression, MinusExpression};
+use crate::ast::Expression;
+use crate::ast::e_minus::{PrefixMinusExpression, MinusExpression};
 
 pub struct MinusParselet {}
