+++ src/parser/p_o_multiplication.rs
@@ -1,8 +1,8 @@
 use crate::parser::{Parselet, ParseError, parse_expression};
 use crate::lexer::Lexer;
-use crate::astplus::Expression;
+use crate::ast::Expression;
 use std::rc::Rc;
-use crate::astplus::e_multiplication::MultiplicationExpression;
+use crate::ast::e_multiplication::MultiplicationExpression;
 
 pub struct MultiplicationParselet {}
