+++ src/parser/p_o_plus.rs
@@ -1,7 +1,7 @@
 use crate::parser::{Parselet, ParseError, parse_expression};
 use crate::lexer::Lexer;
-use crate::astplus::Expression;
-use crate::astplus::e_plus::PlusExpression;
+use crate::ast::Expression;
+use crate::ast::e_plus::PlusExpression;
 use std::rc::Rc;
