+++ src/parser/p_s_let.rs
@@ -1,8 +1,8 @@
 use crate::parser::{Parselet, ParseError, parse_expression};
 use crate::lexer::Lexer;
-use crate::astplus::Expression;
+use crate::ast::Expression;
 use std::rc::Rc;
-use crate::astplus::s_let::LetStatement;
+use crate::ast::s_let::LetStatement;
 
 
 pub struct LetParselet {}
