+++ src/parser/p_v_integer.rs
@@ -1,7 +1,7 @@
 use crate::parser::{Parselet, ParseError, parse_expression};
 use crate::lexer::Lexer;
-use crate::astplus::v_integer::{IntegerExpression};
-use crate::astplus::Expression;
+use crate::ast::v_integer::{IntegerExpression};
+use crate::ast::Expression;
 use std::rc::Rc;
 
 pub struct IntegerParselet {
